Requesting weather from selected and user location

+++ JPMorganChaseWeatherApp/JPMorganChaseWeatherApp/Models/Geocode.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct Geocode: Codable, Hashable {
     let name: String
-    let lat: Float
-    let lon: Float
+    let lat: Double
+    let lon: Double
     let country: String
     let state: String
 }
